scripts/requirements.txt
supabase
requests
python-dotenv # To load environment variables from .env file
numpy         # For numerical operations (vector math) 